Refactor Search resolver
