+++ Makefile
@@ -6,9 +6,11 @@ default::
 help:
 	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | sort | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-25s\033[0m %s\n", $$1, $$2}'
 
-
 setup: ## Generate ssl certificate for development.
 	cd sslcert; openssl req -x509 -nodes -days 365 -newkey rsa:2048 -keyout tls.key -out tls.crt -config req.conf -extensions 'v3_req'
+
+gqlgen: ## Generate graphql model. See: https://gqlgen.com/
+	go run github.com/99designs/gqlgen generate
 
 run: ## Run the service locally.
 	go run main.go playground -v=9
