+++ gqlgen.yml
@@ -54,3 +54,7 @@ models:
       - github.com/99designs/gqlgen/graphql.Int
       - github.com/99designs/gqlgen/graphql.Int64
       - github.com/99designs/gqlgen/graphql.Int32
+  SearchResult:
+    model: github.com/open-cluster-management/search-v2-api/graph/model.SearchResult
+  SearchRelatedResult:
+    model: github.com/open-cluster-management/search-v2-api/graph/model.SearchRelatedResult
