+++ tools.go
@@ -1,0 +1,6 @@
+//go:build tools
+// +build tools
+
+package tools
+
+import _ "github.com/99designs/gqlgen"
